Add files via upload
Find the difference between the total number of CITY entries in the table and the number of distinct CITY entries in the table STATION.
For example, if there are three records in the table with CITY values 'New York', 'New York', and 'Bengalaru', there are 2 different city names: 'New York' and 'Bengalaru'. The query returns because 
Total Number of Records - Total Number of Unique CITY names = 3 - 2 = 1.